Refactored AnonymizedStrategy into strategies package
